Rocky Planet class now supports settlements
There are now methods on the Rocky Planet class that allow
settlements to be added to it. There is currently no way to
remove settlements that are already on the planet. This will
likely be implemented later with a method on the settlement object
that calls for it to be destroyed and dropped from the database.