Added ResourceContainer and ResourceTransform classes
This is the basis of the streaming economy. Settlements and ships
are Resource Containers, which have a set amount of resources and
reference a number of Resource Transform objects, which are
buildings, engines, etc. These Transform objects allow the
container object to modify its resource amounts based on algorithms
that the Transform objects specify.